--- app/data/postgres_storage/__init__.py
@@ -1,5 +1,0 @@
-from app.data.postgres_storage.postgres_storage import (
-    PgStorage,
-    PgStorageConfig,
-    PgStorageConfigSchema,
-)
