+++ app/data/repositories/layer0_repository.py
@@ -3,12 +3,13 @@ from typing import Any, final
 import psycopg
 import structlog
 
-from app.data import interface, postgres_storage, template
+from app.data import interface, template
+from app.lib.storage import postgres
 
 
 @final
 class Layer0Repository(interface.Layer0Repository):
-    def __init__(self, storage: postgres_storage.PgStorage, logger: structlog.BoundLogger) -> None:
+    def __init__(self, storage: postgres.PgStorage, logger: structlog.BoundLogger) -> None:
         self._logger = logger
         self._storage = storage
