+++ app/data/repositories/layer1_repository.py
@@ -3,12 +3,13 @@ from typing import final
 import psycopg
 import structlog
 
-from app.data import interface, model, postgres_storage, template
+from app.data import interface, model, template
+from app.lib.storage import postgres
 
 
 @final
 class Layer1Repository(interface.Layer1Repository):
-    def __init__(self, storage: postgres_storage.PgStorage, logger: structlog.BoundLogger) -> None:
+    def __init__(self, storage: postgres.PgStorage, logger: structlog.BoundLogger) -> None:
         self._logger = logger
         self._storage = storage
