+++ app/lib/queue/task_status.py
@@ -1,0 +1,8 @@
+import enum
+
+
+class TaskStatus(enum.Enum):
+    NEW = "new"
+    IN_PROGRESS = "in_progress"
+    FAILED = "failed"
+    DONE = "done"
